backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
backoff==2.2.1
bcrypt==4.1.3
black==25.9.0
boto3==1.40.50
botocore==1.40.50
build==1.3.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
chromadb==1.2.0
click==8.3.0
coloredlogs==15.0.1
cryptography==46.0.2
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.8.0
durationpy==0.10
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.110.1
filelock==3.20.0
filetype==1.2.0
flake8==7.3.0
flatbuffers==25.9.23
frozenlist==1.8.0
fsspec==2025.9.0
google-ai-generativelanguage==0.8.0
google-api-core==2.26.0
google-auth==2.41.1
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
isort==6.1.0
jmespath==1.0.1
jq==1.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
langchain==1.0.0
langchain-classic==1.0.0
langchain-community==0.4
langchain-core==1.0.0
langchain-google-genai==3.0.0
langchain-text-splitters==1.0.0
langgraph==1.0.0
langgraph-checkpoint==2.1.2
langgraph-prebuilt==1.0.0
langgraph-sdk==0.2.9
langsmith==0.4.37
markdown-it-py==4.0.0
marshmallow==3.26.1
mccabe==0.7.0
mdurl==0.1.2
mmh3==5.2.0
motor==3.3.1
mpmath==1.3.0
multidict==6.7.0
mypy==1.18.2
mypy_extensions==1.1.0
networkx==3.5
numpy==2.3.3
oauthlib==3.3.1
onnxruntime==1.23.1
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.3
ormsgpack==1.11.0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
posthog==5.4.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.0
pydantic-settings==2.11.0
pydantic_core==2.41.1
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.5.0
pypdf==6.1.2
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytokens==0.1.10
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.14.0
s5cmd==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.37.2
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.22.1
tqdm==4.67.1
typer==0.19.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.25.0
uvloop==0.22.1
watchfiles==1.1.0
websocket-client==1.9.0
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
